+++ R/zzz.R
@@ -5,7 +5,8 @@
   op <- options()
   op_duckspatial <- list(
     duckspatial.output_type = "duckspatial_df",
-    duckspatial.max_rows = 1e6
+    duckspatial.max_rows = 1e6,
+    duckspatial.duckdb_storage_version = "v1.5.0"
   )
   toset <- !(names(op_duckspatial) %in% names(op))
   if (any(toset)) options(op_duckspatial[toset])